Changed the units from int to double.

+++ app/src/main/java/com/djacoronel/gwacalculator/model/Course.kt
@@ -2,6 +2,6 @@ package com.djacoronel.gwacalculator.model
 
 class Course(var id: Int,
              val courseCode: String,
-             val units: Int,
+             val units: Double,
              var grade: Double,
              val semester: String)
